snapshot of project "ncurses", label v6_6_20260404

+++ ncurses/curses.priv.h
@@ -35,7 +35,7 @@
  ****************************************************************************/
 
 /*
- * $Id: curses.priv.h,v 1.746 2026/03/28 19:24:13 tom Exp $
+ * $Id: curses.priv.h,v 1.747 2026/04/04 22:09:24 tom Exp $
  *
  *	curses.priv.h
  *
@@ -2609,12 +2609,12 @@ extern NCURSES_EXPORT(int)    _nc_console_checkmintty(int fd, LPHANDLE pMinTTY);
 
 #if USE_TERM_DRIVER && defined(USE_WIN32CON_DRIVER)
 # define NC_ISATTY(fd) (0 != _nc_console_isatty(fd))
-# define NC_READ(fd, buf, count)	read(fd, buf, count)
+# define NC_READ(fd, buf, count)	read(fd, buf, (unsigned)(count))
 #else
 # if defined(_NC_WINDOWS_NATIVE)
-#  define NC_READ(fd, buf, count)	WINCONPTY.read(fd,buf,count)
+#  define NC_READ(fd, buf, count)	WINCONPTY.read(fd, buf, (unsigned)(count))
 # else
-#  define NC_READ(fd, buf, count)	read(fd, buf, count)
+#  define NC_READ(fd, buf, count)	read(fd, buf, (size_t)(count))
 # endif
 # define NC_ISATTY(fd)			isatty(fd)
 #endif
